requirements.txt
# ============================================
# Smart Document Analyst — Dependencies
# Team: Benmouma Salma, Gassi Oumaima
# UIR S8 — Multi-Agent AI Project
# ============================================

# --- Agent Framework ---
crewai>=0.30.0
crewai-tools>=0.4.0

# --- Deep Learning ---
torch>=2.0.0
torchvision>=0.15.0

# --- LLM Backend ---
google-generativeai>=0.5.0

# --- Document Processing ---
Pillow>=10.0.0
pymupdf>=1.24.0
pytesseract>=0.3.10

# --- Report Generation ---
fpdf2>=2.7.0

# --- Data Science & Visualization ---
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
numpy>=1.24.0
tqdm>=4.65.0

# --- Utilities ---
python-dotenv>=1.0.0
pyyaml>=6.0.0

# --- Testing ---
pytest>=7.4.0
